Break persister.py into a package in preparation for adding sqlite persister
